added boss

+++ Game.java
@@ -102,7 +102,7 @@ public class Game extends Canvas implements Runnable {
     }
 
     //clamp method
-    public static int clamp(int var, int min, int max) {
+    public static float clamp(float var, int min, int max) {
         if (var >= max)
             return var = max;
         else if (var <= min)
