+++ GameObject.java
@@ -6,7 +6,7 @@ public abstract class GameObject {
 
     protected int x, y;
     protected ID id;
-    protected int velX, velY;
+    protected float velX, velY;
 
     public GameObject(int x, int y, ID id){
         this.x = x;
@@ -50,11 +50,11 @@ public abstract class GameObject {
         this.velY = velY;
     }
 
-    public int getVelX() {
+    public float getVelX() {
         return velX;
     }
 
-    public int getVelY() {
+    public float getVelY() {
         return velY;
     }
 }
