+++ HUD.java
@@ -4,7 +4,7 @@ import java.awt.*;
 
 public class HUD {
 
-    public static int HEALTH = 100;
+    public static float HEALTH = 100;
     private int greenValue = 255;
 
     private int score = 0;
@@ -13,8 +13,8 @@ public class HUD {
     public void tick() {
         HEALTH = Game.clamp(HEALTH, 0, 100);
 
-        greenValue = Game.clamp(greenValue, 0, 255);
-        greenValue = HEALTH * 2;
+        greenValue = (int)Game.clamp((float)greenValue, 0, 255);
+        greenValue = (int)HEALTH * 2;
 
         score++;
     }
@@ -23,7 +23,7 @@ public class HUD {
         g.setColor(Color.gray);
         g.fillRect(15, 15, 200, 32);
         g.setColor(new Color(75, greenValue, 0));
-        g.fillRect(15, 15, HEALTH * 2, 32);
+        g.fillRect(15, 15, (int)HEALTH * 2, 32);
         g.setColor(Color.white);
         g.drawRect(15, 15, 200, 32);
